Refactor the standardized JSON encoder out of the db.datatypes library
First written for the database layer and still living there to avoid
premature abstraction, but with the advent of non-database code now
wanting to use it, it's time to move out.
